Вроде фикс нулевых уровня моба и макс.хп, на которых код чертыхается в разных местах. Щас у залоаденного моба не меньше 1 должны быть оба.

+++ src/db.h
@@ -19,7 +19,7 @@
 #include "pugixml.hpp"
 
 // Максимальный уровень мобов
-const int MAX_MOB_LEVEL = 50;
+const short MAX_MOB_LEVEL = 50;
 
 // arbitrary constants used by index_boot() (must be unique)
 #define MAX_PROTO_NUMBER 999999	//Максимально возможный номер комнаты, предмета и т.д.
